+++ ibuffer suite/ibufplayer~/ibufplayer~.cpp
@@ -171,7 +171,7 @@ void *ibufplayer_new(t_symbol *s, long argc, t_atom *argv)
     
     // Arguments
     
-    long non_attr_argc = attr_args_offset(argc, argv);
+    long non_attr_argc = attr_args_offset(static_cast<short>(argc), argv);
     
     t_atom_long obj_n_chans = non_attr_argc > 0 ? atom_getlong(argv + 0) : 1;
     obj_n_chans = std::min(static_cast<t_atom_long>(max_num_chans), std::max(static_cast<t_atom_long>(1), obj_n_chans));
@@ -209,7 +209,7 @@ void *ibufplayer_new(t_symbol *s, long argc, t_atom *argv)
     
     // Set attributes from arguments
     
-    attr_args_process(x, argc, argv);
+    attr_args_process(x, static_cast<short>(argc), argv);
     
     return x;
 }
